# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "仿百度网盘的后端系统，支持文件存储、分享、收藏等功能",
        "title": "云盘系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/shares": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "获取分享列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ShareListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/shares/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "获取分享详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分享ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ShareResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "更新分享",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分享ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ShareUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "取消分享",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分享ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/s/{token}": {
            "get": {
                "description": "通过token访问分享的文件信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "访问分享",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享令牌",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问密码",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ShareDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "密码错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "分享不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/s/{token}/download": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "下载分享文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享令牌",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问密码",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到文件",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shares": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建文件分享链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享管理"
                ],
                "summary": "创建分享",
                "parameters": [
                    {
                        "description": "分享信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ShareCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ShareResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "文件不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.ShareCreateRequest": {
            "type": "object",
            "required": [
                "file_id"
            ],
            "properties": {
                "expire_days": {
                    "description": "过期天数",
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                },
                "file_id": {
                    "type": "string"
                },
                "max_downloads": {
                    "description": "最大下载次数",
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "password": {
                    "description": "访问密码",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                }
            }
        },
        "request.ShareUpdateRequest": {
            "type": "object",
            "properties": {
                "expire_days": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                },
                "max_downloads": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
...
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.FileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_dir": {
                    "type": "boolean"
                },
                "mime_type": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "size_text": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ShareDetailResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "download_left": {
                    "type": "integer"
                },
                "expire_time": {
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/response.FileResponse"
                },
                "need_password": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "response.ShareListResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "shares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ShareResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.ShareResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer"
                },
                "expire_time": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_size_text": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_downloads": {
                    "type": "integer"
                },
                "share_link": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}